Converted memoizeFetch to cachePromise
After considering this a bit more it felt like memoization was overkill
and perhaps deterimental as it maintained a reference to all previous
promises. By overriding the hashFunction the way we were it was
essentially converting the memoize function into a single-element cache.

This cachePromise implementation makes that more obvious.